Change name to PendingTasksCount
