+++ rest/rest-cli/src/main/java/org/ow2/proactive_grid_cloud_portal/cli/cmd/rm/RmStatsCommand.java
@@ -49,7 +49,7 @@ public class RmStatsCommand extends AbstractCommand implements Command {
                                                                      "MaxFreeNodes", "AverageInactivity", "Status",
                                                                      "DownNodesCount", "AverageActivity",
                                                                      "MaxDownNodes", "ConfiguringNodesCount",
-                                                                     "MaxToBeReleasedNodes", "NumberPendingTasks" };
+                                                                     "MaxToBeReleasedNodes", "PendingTasksCount" };
 
     @Override
     public void execute(ApplicationContext currentContext) throws CLIException {
