+++ rm/rm-client/src/main/java/org/ow2/proactive/resourcemanager/common/util/RMProxyUserInterface.java
@@ -446,8 +446,8 @@ public class RMProxyUserInterface extends RMListenerProxy implements ResourceMan
     }
 
     @Override
-    public void setTotalNumberOfPendingTasks(int totalNumberOfPendingTasks) {
-        this.target.setTotalNumberOfPendingTasks(totalNumberOfPendingTasks);
+    public void setPendingTasksCount(int pendingTasksCount) {
+        this.target.setPendingTasksCount(pendingTasksCount);
     }
 
     @Override
