+++ rm/rm-client/src/main/java/org/ow2/proactive/resourcemanager/frontend/ResourceManager.java
@@ -557,5 +557,5 @@ public interface ResourceManager {
      */
     boolean areNodesRecoverable(NodeSet nodes);
 
-    void setTotalNumberOfPendingTasks(int totalNumberOfPendingTasks);
+    void setPendingTasksCount(int pendingTasksCount);
 }
