+++ rm/rm-server/src/main/java/org/ow2/proactive/resourcemanager/core/jmx/mbean/RuntimeDataMBean.java
@@ -68,7 +68,7 @@ public interface RuntimeDataMBean {
      */
     int getAvailableNodesCount();
 
-    int getNumberPendingTasks();
+    int getPendingTasksCount();
 
     /**
      * Returns the current number of nodes in {@link NodeState#CONFIGURING} state.
