+++ rm/rm-server/src/main/java/org/ow2/proactive/resourcemanager/frontend/RMMonitoringImpl.java
@@ -42,7 +42,6 @@ import org.objectweb.proactive.core.UniqueID;
 import org.objectweb.proactive.core.body.request.Request;
 import org.objectweb.proactive.core.util.wrapper.BooleanWrapper;
 import org.objectweb.proactive.extensions.annotation.ActiveObject;
-import org.ow2.proactive.PendingTasksListener;
 import org.ow2.proactive.resourcemanager.authentication.Client;
 import org.ow2.proactive.resourcemanager.common.RMConstants;
 import org.ow2.proactive.resourcemanager.common.event.*;
@@ -69,7 +68,7 @@ import org.ow2.proactive.resourcemanager.utils.AtomicRMStatisticsHolder;
  * @since ProActive Scheduling 0.9
  */
 @ActiveObject
-public class RMMonitoringImpl implements RMMonitoring, RMEventListener, InitActive, RunActive, PendingTasksListener {
+public class RMMonitoringImpl implements RMMonitoring, RMEventListener, InitActive, RunActive {
     private static final Logger logger = Logger.getLogger(RMMonitoringImpl.class);
 
     // Attributes
@@ -419,9 +418,8 @@ public class RMMonitoringImpl implements RMMonitoring, RMEventListener, InitActi
         queueEvent(event);
     }
 
-    @Override
-    public void notifyPendingTasksCount(int count) {
-        RMMonitoringImpl.rmStatistics.notifyPendingTasksCount(count);
+    public void setPendingTasksCount(int count) {
+        RMMonitoringImpl.rmStatistics.setPendingTasksCount(count);
     }
 
     /**
